[Matrix]: Fixing data typing problems

+++ flash-graph/matrix/kmeans.h
@@ -108,7 +108,7 @@ namespace fg {
  **/
 unsigned compute_kmeans(const double* matrix, double* clusters,
 		unsigned* cluster_assignments, unsigned* cluster_assignment_counts,
-		const size_t num_rows, const size_t num_cols, const unsigned k,
+		const unsigned num_rows, const unsigned num_cols, const unsigned k,
 		const unsigned MAX_ITERS, const int max_threads,
         const std::string init="kmeanspp", const double tolerance=-1,
         const std::string dist_type="eucl");
@@ -116,7 +116,7 @@ unsigned compute_kmeans(const double* matrix, double* clusters,
 /** See `compute_kmeans` for argument list */
 unsigned compute_min_kmeans(const double* matrix, double* clusters_ptr,
         unsigned* cluster_assignments, unsigned* cluster_assignment_counts,
-		const size_t num_rows, const size_t num_cols, const unsigned k,
+		const unsigned num_rows, const unsigned num_cols, const unsigned k,
         const unsigned MAX_ITERS, const int max_threads,
         const std::string init="kmeanspp", const double tolerance=-1,
         const std::string dist_type="eucl");
